+++ docs/layout/header.vue
@@ -2,8 +2,8 @@
   <header class="header">
     <div class="header-container">
       <router-link :to="{ path: '/' }" class="logo-block">
-        <i class="wot-design-logo"></i>
-        <span>Wot Design Mini</span>
+        <img class="wot-design-logo" src="../assets/img/wot-design.png" alt="wot design mini" />
+        <h1 class="wot-design-title">京麦 Wot Design Mini</h1>
       </router-link>
       
       <ul class="header-tab">
@@ -23,15 +23,23 @@
         </li>
         <!-- 版本控制 -->
         <li class="header-tab__item version-control" v-show="isComponentPage" >
-          <a href="#" class="header-tab__link " @click="showOption">{{ version }}</a>
-          <div class="wot-dropdown" v-show="isShowOption">
+          <a class="header-tab__link header-tab__with-arrow" @click="showOption">{{ version }}</a>
+          <transition name="drop-scale-in">
+            <div class="wot-dropdown" v-show="isShowOption">
               <ul class="wot-dropdown-menu">
-                <li class="wot-dropdown-item" v-for="item in versions" :value="item" :key="item" @click="switchVersion(item)">
+                <li
+                  class="wot-dropdown-item"
+                  :class="{ 'is-active': item === version }"
+                  v-for="item in formatVersions"
+                  :key="item"
+                  @click="switchVersion(item)"
+                >
                   {{ item }}
                 </li>
               </ul>
-              <i class="popper__arrow" ></i>
-          </div>
+              <i class="popper__arrow"></i>
+            </div>
+          </transition>
         </li>
       </ul>
     </div>
@@ -40,8 +48,8 @@
 
 <script>
 import pagesConfig from '../pages.config.json'
-import versions from '../versions.json'
 import search from './search'
+import axios from 'axios'
 const { version } = require('../../package.json')
 
 export default {
@@ -49,17 +57,29 @@ export default {
   data () {
     return {
       pages: pagesConfig,
-      versions: this.filter(),
+      versions: [],
       isComponentPage: true,
       isShowOption: false,
       version
+    }
+  },
+  computed: {
+    formatVersions () {
+      return this.filter(this.versions)
     }
   },
   methods: {
     showOption () {
       this.isShowOption = !this.isShowOption
     },
-    filter () {
+    getVersions () {
+      const requestUrl = process.env.NODE_ENV === 'dev' ? '/static/public/versions.json' : '/wot-design-mini/static/public/versions.json'
+
+      axios.get(requestUrl).then(res => {
+        this.versions = res.data
+      })
+    },
+    filter (versions) {
       const keys = versions
       let result = []
       let preVersionList = []
@@ -85,7 +105,7 @@ export default {
           })
         }
       })
-      return result.sort()
+      return result.sort((a, b) => a - b)
     },
     switchVersion (selected) {
       this.isShowOption = !this.isShowOption
@@ -93,6 +113,27 @@ export default {
       // location.hash
       window.location.href = `${ location.origin }/wot-design-mini/${ selected }/#/components/introduction`
     },
+    clickOutside (event) {
+      let clickDom = event.target
+
+      // 监听常见问题标题点击
+      while (clickDom.className.indexOf('version-control') === -1 && clickDom !== document.body) {
+        clickDom = clickDom.parentNode
+      }
+
+      if (clickDom.className.indexOf('version-control') === -1) {
+        this.isShowOption = false
+      }
+    }
+  },
+  mounted () {
+    document.body.addEventListener('click', this.clickOutside)
+  },
+  beforeDestroy () {
+    document.body.removeEventListener('click', this.clickOutside)
+  },
+  created () {
+    this.getVersions()
   }
 }
 </script>
@@ -108,7 +149,7 @@ export default {
   top: 0;
   left: 0;
   width: 100%;
-  border-bottom: 1px solid #eee;
+  box-shadow: 0 2px 10px #eee;
   background: $color-bg;
   z-index: 100;
 }
@@ -117,6 +158,7 @@ export default {
   height: $layout-header-height;
   line-height: $layout-header-height;
 }
+
 .logo-block {
   display: inline-block;
   height: $layout-header-height;
@@ -134,9 +176,14 @@ export default {
   margin-right: 10px;
   width: 40px;
   height: 40px;
-  background: url('../assets/img/wot-design.png') no-repeat;
-  background-size: cover;
   vertical-align: middle;
+}
+.wot-design-title {
+  display: inline-block;
+  font-weight: normal;
+  font-size: 18px;
+  color: $color-theme;
+  margin: 0;
 }
 .header-tab {
   float: right;
@@ -153,25 +200,31 @@ export default {
   font-size: $fs-title;
   line-height: 25px;
   color: $color-text-light;
+  transition: color .3s;
+  cursor: pointer;
+  user-select: none;
 
   &:hover {
     color: $color-important;
   }
+}
+.header-tab__with-arrow::after {
+  position: absolute;
+  display: inline-block;
+  content: '';
+  width: 0;
+  height: 0;
+  top: 50%;
+  margin-top: -3px;
+  right: 0;
+  border: 6px solid rgba(0,0,0,0);
+  border-top-color: #ccc;
 }
 .header-tab__link--active{
   color: $color-theme;
 
   &:hover {
     color: $color-theme;
-  }
-  &::after {
-    position: absolute;
-    content: '';
-    width: 70%;
-    height: 2px;
-    left: 15%;
-    bottom: 0;
-    background: $color-theme;
   }
 }
 .version-control{
@@ -184,33 +237,59 @@ export default {
     background-color: #ebebeb;
   }
 }
+.wot-dropdown {
+  position: relative;
+}
 .wot-dropdown-menu{
-  padding: 10px 5px;
-  margin: 0 10px;
+  position: absolute;
+  right: 0;
+  top: 100%;
+  padding: 10px 0;
+  width: 90px;
+  max-height: 300px;
+  overflow: auto;
   border: none;
   background-color: transparent;
   font-size: 14px;
   color: #666;
-  position: absolute;
   background-color: #fff;
   border: 1px solid #ebeef5;
   border-radius: 2px;
-  box-shadow: 0 2px 12px 0 rgba(0,0,0,.1);
+  box-shadow: 0 2px 12px 4px rgba(0,0,0,.1);
+  scrollbar-width: thin;
+  scrollbar-color: rgba(0, 0, 0, 0.3) #fefefe;
+
+  &::-webkit-scrollbar-track-piece {
+    background-color: transparent;
+  }
+  &::-webkit-scrollbar {
+    width: 6px;
+    height: 6px;
+  }
+  &::-webkit-scrollbar-thumb {
+    width: 6px;
+    height: 6px;
+    opacity: 0.5;
+    border-radius: 3px;
+    background-color: rgba(0, 0, 0, 0.3);
+  }
 }
 .wot-dropdown-item{
   list-style: none;
   height: 30px;
-  line-height: 32px;
-  padding: 0 10px;
+  line-height: 30px;
+  padding-left: 15px;
   margin: 0;
   font-size: 12px;
   color: #464c5b;
   cursor: pointer;
   font-size: 14px;
   outline: none;
-  &:hover{
-    background-color: #f5f7f9;
-    color: #464c5b;
+  transition: background .3s, color .3s;
+
+  &:hover, &.is-active, &.is-active:hover {
+    background-color: mix(#0083ff, #fff, 10%);
+    color: #0083ff;
   }
 }
 .popper__arrow{
@@ -219,8 +298,32 @@ export default {
   position: absolute;
   left: 50%;
   bottom: -2px;
+  margin-left: -3px;
   border-width: 6px;
   border-color: transparent;
   border-bottom-color: white;
+}
+.drop-scale-in-enter, .drop-scale-in-leave-to {
+  transform: scaleY(0);
+}
+.drop-scale-in-enter-active, .drop-scale-in-leave-active {
+  transition: transform .2s;
+}
+@media (max-width: 1366px) {
+  .header-container {
+    margin: 0 30px;
+  }
+}
+@media (max-width: 773px) {
+  .wot-design-title {
+    font-size: 16px;
+  }
+  .header-container {
+    margin: 0 15px;
+  }
+  .header-tab__link {
+    padding: 6px 12px;
+    font-size: 14px;
+  }
 }
 </style>