+++ example/pages/datetimePicker/index.js
@@ -1,3 +1,5 @@
+import Toast from '../../dist/toast/toast.js'
+
 Page({
   data: {
     value1: Date.now(),
@@ -5,6 +7,10 @@ Page({
     value3: '09:20',
     value4: Date.now(),
     value5: Date.now(),
+    value6: Date.now(),
+    value7: Date.now(),
+    value8: Date.now(),
+    value9: Date.now(),
     formatter (type, value) {
       switch (type) {
       case 'year':
@@ -29,6 +35,22 @@ Page({
     },
     displayFormat (items) {
       return `${items[0].label}年${items[1].label}月${items[2].label}日 ${items[3].label}:${items[4].label}`
+    },
+    beforeConfirm (value, resolve, picker) {
+      picker.setData({
+        loading: true
+      })
+      setTimeout(() => {
+        picker.setData({
+          loading: false
+        })
+        if (value > Date.now()) {
+          resolve(false)
+          Toast.error('不能选择大于今天的日期')
+        } else {
+          resolve(true)
+        }
+      }, 2000)
     }
   },
   /** picker触发confirm事件，同步触发confirm事件 */
@@ -36,5 +58,10 @@ Page({
   },
   /** picker触发cancel事件，同步触发cancel事件 */
   onCancel () {
+  },
+  handleConfirm ({ detail }) {
+    this.setData({
+      value6: detail
+    })
   }
 })