+++ packages/mixins/datetimePickerView.js
@@ -8,7 +8,15 @@ const isValidDate = date => isDef(date) && !Number.isNaN(date)
 /** @description 保证num不超出min和max的范围 */
 const range = (num, min, max) => Math.min(Math.max(num, min), max)
 /** @description 不满10补0 */
-const padZero = val => `00${val}`.slice(-2)
+const padZero = (number, length = 2) => {
+  number = number + ''
+
+  while (number.length < length) {
+    number = '0' + number
+  }
+
+  return number
+}
 /**
  * @description 生成n个元素，并使用iterator接口进行填充
  * @param n
@@ -22,18 +30,6 @@ const times = (n, iteratee) => {
     result[index] = iteratee(index)
   }
   return result
-}
-/**
- * @description 还原数据，例如 2019年 ->getTrueValue-> 2019
- * @param {String} formattedValue
- * @return {Number}
- */
-const getTrueValue = (formattedValue) => {
-  if (!formattedValue) return
-  while (isNaN(parseInt(formattedValue, 10))) {
-    formattedValue = formattedValue.slice(1)
-  }
-  return parseInt(formattedValue, 10)
 }
 /**
  * @description 获取某年某月有多少天
@@ -178,7 +174,10 @@ export default function () {
       updateColumns () {
         const { formatter = defaultFormatter } = this.data
         return this.getOriginColumns().map(column => {
-          return column.values.map(value => formatter(column.type, value))
+          return column.values.map(value => ({
+            label: formatter(column.type, padZero(value)),
+            value
+          }))
         })
       },
       /**
@@ -189,9 +188,7 @@ export default function () {
         const { filter } = this.data
         return this.getRanges().map(({ type, range }) => {
           let values = times(range[1] - range[0] + 1, index => {
-            let value = range[0] + index
-            value = type === 'year' ? `${value}` : padZero(value)
-            return value
+            return range[0] + index
           })
 
           if (filter) {
@@ -345,28 +342,18 @@ export default function () {
        */
       updateColumnValue (value) {
         const values = []
-        const { type, formatter = defaultFormatter } = this.data
+        const { type } = this.data
         const date = new Date(value)
 
         if (type === 'time') {
           const pair = value.split(':')
-          values.push(
-            formatter('hour', pair[0]),
-            formatter('minute', pair[1])
-          )
+          values.push(pair[0], pair[1])
         } else {
-          values.push(
-            formatter('year', `${date.getFullYear()}`),
-            formatter('month', padZero(date.getMonth() + 1))
-          )
+          values.push(date.getFullYear(), date.getMonth() + 1)
           if (type === 'date') {
-            values.push(formatter('date', padZero(date.getDate())))
+            values.push(date.getDate())
           } else if (type === 'datetime') {
-            values.push(
-              formatter('date', padZero(date.getDate())),
-              formatter('hour', padZero(date.getHours())),
-              formatter('minute', padZero(date.getMinutes()))
-            )
+            values.push(date.getDate(), date.getHours(), date.getMinutes())
           }
         }
         this.setData({ innerValue: value })
@@ -390,11 +377,11 @@ export default function () {
           return
         }
         /** 重新计算年月日时分秒，修正时间。 */
-        const values = picker.getLabels()
-        const year = getTrueValue(values[0])
-        const month = getTrueValue(values[1])
+        const values = picker.getValues()
+        const year = values[0]
+        const month = values[1]
         const maxDate = getMonthEndDay(year, month)
-        let date = getTrueValue(values[2])
+        let date = values[2]
         if (type === 'year-month') {
           date = 1
         }
@@ -402,8 +389,8 @@ export default function () {
         let hour = 0
         let minute = 0
         if (type === 'datetime') {
-          hour = getTrueValue(values[3])
-          minute = getTrueValue(values[4])
+          hour = values[3]
+          minute = values[4]
         }
         const value = new Date(year, month - 1, date, hour, minute)
         /** 根据计算选中项的时间戳，重新计算所有的选项列表 */